First cut of custom clipboard paste types (mac only)
This checkpoint attempts to generalize the idea of pasting custom
types on macOS in a way that I *hope* is fairly flexible.

I don't expect this to last, because I would like to have a common
mechanism for reading and writing of custom types to the clipboard,
but I'd like to check this in before I make too much of a mess.