Impl most of a new clipboard API
This is the bones of a more flexible system for copy/paste. The most
basic idea is that when the user wants to copy or paste custom types,
they define some marker struct and implement the `ClipboardRead` and or
`ClipboardWrite` traits.

Methods on these traits can be implemented per-platform, to return
platform specific information about how to serialize and deserialize
a given type. If impls are omitted for a platform, the type is
ignored on that platform.

The intent with this API is to be maxmimally accomodating to the fact
that copy/paste APIs are very different on different platforms;
instead of trying to come up with a lowest common feature set,
we instead try to be as general as possible, without sacrificing
ergonomics.

TODOs:
- I would like to prove this out on at least one additional platform
- I would like to make copy/cut easier to use with this API; in
particular I would like to support copying multiple data-types.